requirements.txt
flask>=3.0
werkzeug>=3.0
python-dotenv>=1.0
flask-sqlalchemy>=3.1
flask-migrate>=4.0
flask-login>=0.6
